+++ bascomtask-core/src/main/java/com/ebay/bascomtask/core/TaskInterface.java
@@ -16,43 +16,52 @@
  **************************************************************************/
 package com.ebay.bascomtask.core;
 
+import com.ebay.bascomtask.exceptions.MisplacedTaskMethodException;
+
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
 
 /**
- * Common interface that all user POJO task object interfaces should implement, like this:
+ * Common interface that all user POJO task interfaces must implement in order to be added to an Orchestrator,
+ * for example like this:
  * <pre>{@code
  * interface IDelay implements TaskInterface<IDelay> {
- *    public void exec(...);
+ *    // Add test methods here
  * }
  *
  * class Delay implements IDelay {...}
  * }</pre>
- * <p>
- * Note that Delay or IDelay can implement/extend other interfaces, but among all of them only
- * one should be designated as a TaskInterface in the manner above.
- * <p>
- * There are no methods to implement from this interface, but it does provide several common
- * methods on tasks. It also allows the BT task-dependency mechanism to work properly.
  *
- * @param <T> identifies that interface as one that defines task methods
+ * <p>Task classes such as <code>Delay</code> or its interface <code>IDelay</code> above can implement/extend other
+ * interfaces, but among all of them only one should be designated as argument to TaskInterface in the manner above.
+ *
+ * <p>There are no required methods to implement from this interface since all have default implementations. For
+ * uncommon cases some of these methods may usefully be overridden as noted in each javadoc. Several convenience
+ * methods are also provided for wrapping and unwrapping values in/from CompletableFutures. There use is completely
+ * optional.
+ *
+ * @param <T> identifies that interface that defines task methods
  * @author Brendan McCarthy
+ * @see Orchestrator#task(TaskInterface) 
  */
 public interface TaskInterface<T> {
 
     /**
-     * Provides a name for this task, useful for logging and profiling. Subclasses can optionally override
-     * if they want to allow setting a default name. When invoked on a TaskWrapper (the object that is returned by
-     * adding a task to an Orchestrator), it does not pass that call to this method. TaskWrappers manages names
-     * on their own (any number of TaskWrappers can reference the same user task).
+     * Provides a name for this task, useful for logging and profiling.
+     *
+     * TaskWrappers (as returned by {@link Orchestrator#task(TaskInterface)}, already support this method which
+     * therefore works in expressions like <pre><code>$.task(myTask).name("myTask").myMethod()</code></pre>.
+     * That is the typical usage, so there is normally no need to override this method, but subclasses can
+     * override if desired to also enable invocation directly on the task, e.g.
+     * <pre><code>$.task(myTask.name("myTask")).myMethod()</code></pre>.
      *
      * @param name to set
      * @return this
      */
     default T name(String name) {
-        throw new RuntimeException("Invalid call for setting name \"" + name + "\", subclass has not overridden this method");
+        throw new MisplacedTaskMethodException(this,"name");
     }
 
     /**
@@ -60,7 +69,7 @@ public interface TaskInterface<T> {
      * override if they want to expose something other than the class name as the default name. Whether or not
      * overridden, any TaskWrapper (the object that is returned by adding a task to an Orchestrator) that wraps
      * this object (there can be more than one) will manage its own name and use this value returned here only
-     * if not explicitly set during task orchestration.
+     * if not explicitly set during task wiring.
      *
      * @return name
      */
@@ -69,30 +78,88 @@ public interface TaskInterface<T> {
     }
 
     /**
-     * Forces execution within the outer thread, i.e. prevents a new thread from being spawned for this task.
-     * Equivalent to using the {@link com.ebay.bascomtask.annotations.Light} annotation on a task method.
+     * Forces execution of task methods on this object to be done inline, i.e. prevents a new thread from being
+     * spawned when an <code>Orchestrator</code> might otherwise do so.
      *
-     * <p>Reverses any previous call to {@link #runSpawned()}
+     * TaskWrappers (as returned by {@link Orchestrator#task(TaskInterface)}, already support this method which
+     * therefore works in expressions like <pre><code>$.task(myTask).light().myMethod()</code></pre>.
+     * That is the typical usage, so there is normally no need to override this method, but subclasses can
+     * override if desired to also enable invocation directly on the task, e.g.
+     * <pre><code>$.task(myTask.light()).myMethod()</code></pre>.
+     *
+     * <p>This call reverses any previous call to {@link #runSpawned()}. The impact of this call is redundant for
+     * task methods that already have {@link com.ebay.bascomtask.annotations.Light} annotation.
      *
      * @return this
      */
-    @SuppressWarnings("unchecked")
     default T light() {
-        return (T) this;
+        throw new MisplacedTaskMethodException(this,"light");
     }
 
     /**
-     * Forces a new thread to be allocated for this task. This can be useful to keep the calling task free
-     * for other purposes. This overrides the default behavior, which is spawn threads for each task &gt; 1 when an
-     * executing thread finds more than one task ready to fire.
+     * Indicates the default weight of task methods that do not have a {@link com.ebay.bascomtask.annotations.Light}
+     * annotation.
      *
-     * <p>Reverses any previous call to {@link #light()} or any {@link com.ebay.bascomtask.annotations.Light} annotation
+     * @return true iff by default a thread should never be spawned for task methods on this class
+     */
+    default boolean isLight() {
+        return false;
+    }
+
+    /**
+     * Forces a new thread to be allocated for task methods on this class, which can be useful to keep the calling
+     * thread free for other purposes. This overrides the default behavior, which is spawn threads
+     * when an executing thread finds more than one task ready to fire.
+     *
+     * TaskWrappers (as returned by {@link Orchestrator#task(TaskInterface)}, already support this method which
+     * therefore works in expressions like <pre><code>$.task(myTask).runSpawned().myMethod()</code></pre>.
+     * That is the typical usage, so there is normally no need to override this method, but subclasses can
+     * override if desired to also enable invocation directly on the task, e.g.
+     * <pre><code>$.task(myTask.runSpawned()).myMethod()</code></pre>.
+     *
+     * <p>This call reverses any previous call to {@link #light()}. Its effect is superseded by a
+     * {@link com.ebay.bascomtask.annotations.Light} annotation on a task method if present.
      *
      * @return this
      */
     @SuppressWarnings("unchecked")
     default T runSpawned() {
-        return (T) this;
+        throw new MisplacedTaskMethodException(this,"runSpawned");
+    }
+
+    /**
+     * Indicates that task methods should have threads spawned for them.
+     *
+     * @return true iff by default a thread should be spawned for task methods on this class
+     */
+    default boolean isRunSpawned() {
+        return false;
+    }
+
+    /**
+     * Forces immediate activation of task methods instead of the default behavior which is that task methods
+     * are only lazily activated according to the rules described in {@link Orchestrator#task(TaskInterface)}.
+     *
+     * TaskWrappers (as returned by {@link Orchestrator#task(TaskInterface)}, already support this method which
+     * therefore works in expressions like <pre><code>$.task(myTask).activate(true).myMethod()</code></pre>.
+     * That is the typical usage, so there is normally no need to override this method, but subclasses can
+     * override if desired to also enable invocation directly on the task, e.g.
+     * <pre><code>$.task(myTask.activate(true)).myMethod()</code></pre>.
+     *
+     * @return this
+     */
+    default T activate() {
+        throw new MisplacedTaskMethodException(this,"activate");
+    }
+
+    /**
+     * Indicates that task methods should be activated immediately. Subclasses can override to change the
+     * default false return value if desired.
+     *
+     * @return true iff by default a thread should never be spawned for task methods on this class
+     */
+    default boolean isActivate() {
+        return false;
     }
 
     /**
@@ -120,8 +187,7 @@ public interface TaskInterface<T> {
     }
 
     /**
-     * Convenience method creates a CompletableFuture from a constant value, which
-     * may or may not be null.
+     * Convenience method creates a CompletableFuture from a constant value, which may or may not be null.
      *
      * @param arg value to wrap
      * @param <R> type of value returned
@@ -136,7 +202,7 @@ public interface TaskInterface<T> {
      * even for logically void methods in order to expose common methods (such as being able to start
      * task execution) on the return result.
      *
-     * @return a void result
+     * @return a Void result
      */
     default CompletableFuture<Void> complete() {
         return CompletableFuture.allOf();