+++ bascomtask-core/src/test/java/com/ebay/bascomtask/core/BascomTaskFutureTest.java
@@ -143,68 +143,15 @@ public class BascomTaskFutureTest extends BaseOrchestratorTest {
 
     @Test
     public void thenAccept() {
-        CompletableFuture<Integer> cf = $.task(task(1)).ret(1);
+        CompletableFuture<Integer> cf = $.task(task(0)).ret(1);
         assertNotNull(cf.thenAccept(v -> hit()));
     }
 
     @Test
-    public void thenAcceptAsync() {
+    public void thenAcceptActivate() {
         CompletableFuture<Integer> cf = $.task(task(1)).ret(1);
-        assertNotNull(cf.thenAcceptAsync(v -> hit()));
-    }
-
-    @Test
-    public void thenAcceptAsyncExecutor() {
-        CompletableFuture<Integer> cf = $.task(task(1)).ret(1);
-        assertNotNull(cf.thenAcceptAsync(v -> hit(), executor));
-    }
-
-
-    @Test
-    public void thenAcceptBoth() {
-        CompletableFuture<Integer> other = $.task(task(1)).ret(5);
-        CompletableFuture<Integer> cf = $.task(task(1)).ret(1);
-        assertNotNull(cf.thenAcceptBoth(other, (x, y) -> hit(6, x + y)));
-    }
-
-    @Test
-    public void thenAcceptBothAsync() {
-        CompletableFuture<Integer> other = $.task(task(1)).ret(5);
-        CompletableFuture<Integer> cf = $.task(task(1)).ret(1);
-        assertNotNull(cf.thenAcceptBothAsync(other, (x, y) -> hit(6, x + y)));
-    }
-
-    @Test
-    public void thenAcceptBothAsyncExecutor() {
-        CompletableFuture<Integer> other = $.task(task(1)).ret(3);
-        CompletableFuture<Integer> cf = $.task(task(1)).ret(1);
-        assertNotNull(cf.thenAcceptBothAsync(other, (x, y) -> hit(4, x + y), executor));
-    }
-
-
-    @Test
-    public void thenRun() {
-        CompletableFuture<Integer> cf = $.task(task(1)).ret(1);
-        assertNotNull(cf.thenRun(this::hit));
-    }
-
-    @Test
-    public void thenRunAsync() {
-        CompletableFuture<Integer> cf = $.task(task(1)).ret(1);
-        assertNotNull(cf.thenRunAsync(this::hit));
-    }
-
-    @Test
-    public void thenRunAsyncExecutor() {
-        CompletableFuture<Integer> cf = $.task(task(1)).ret(1);
-        assertNotNull(cf.thenRunAsync(this::hit, executor));
-    }
-
-
-    @Test
-    public void thenApply() {
-        CompletableFuture<Integer> cf = $.task(task(1)).ret(1);
-        assertNotNull(cf.thenApply(v -> hit()));
+        $.execute(cf);
+        assertNotNull(cf.thenAccept(v -> hit()));
     }
 
     @Test
@@ -212,6 +159,7 @@ public class BascomTaskFutureTest extends BaseOrchestratorTest {
         CompletableFuture<Integer> cf = $.task(task(2)).ret(1);
 
         CompletableFuture<Integer> c2 = cf.thenApply(v -> hit() + v + 10).thenApply(v -> hit() + v + 20);
+        $.execute(cf);
         assertEquals(34, (int) c2.get());
     }
 
@@ -219,41 +167,8 @@ public class BascomTaskFutureTest extends BaseOrchestratorTest {
     public void thenApplyThenAccept() throws Exception {
         CompletableFuture<Integer> cf = $.task(task(2)).ret(1);
         CompletableFuture<Void> c2 = cf.thenApply(v -> hit() + v + 10).thenAccept(v -> hit());
+        $.execute(cf);
         c2.get();
-    }
-
-    @Test
-    public void thenApplyAsync() throws Exception {
-        CompletableFuture<Integer> cf = $.task(task(1)).ret(1);
-        int got = cf.thenApplyAsync(v -> hit()).get();
-        assertEquals(1, got);
-    }
-
-    @Test
-    public void thenApplyAsyncExecutor() throws Exception {
-        CompletableFuture<Integer> cf = $.task(task(1)).ret(1);
-        int got = cf.thenApplyAsync(v -> hit(), executor).get();
-        assertEquals(1, got);
-    }
-
-    @Test
-    public void thenCombine() throws Exception {
-        CountingTask task = task(0);
-        CompletableFuture<Integer> cf1 = $.task(task).ret(1);
-        CompletableFuture<Integer> cf2 = $.task(task).ret(1);
-        CompletableFuture<Integer> combine = cf1.thenCombine(cf2, Integer::sum);
-        int got = combine.get();
-        assertEquals(2, got);
-    }
-
-    @Test
-    public void thenCombineAsync() throws Exception {
-        CountingTask task = task(0);
-        CompletableFuture<Integer> cf1 = $.task(task).ret(1);
-        CompletableFuture<Integer> cf2 = $.task(task).ret(1);
-        CompletableFuture<Integer> combine = cf1.thenCombineAsync(cf2, Integer::sum);
-        int got = combine.get();
-        assertEquals(2, got);
     }
 
     @Test
@@ -262,6 +177,7 @@ public class BascomTaskFutureTest extends BaseOrchestratorTest {
         CompletableFuture<Integer> cf1 = $.task(task).ret(1);
         CompletableFuture<Integer> cf2 = $.task(task).ret(1);
         CompletableFuture<Integer> combine = cf1.thenCombineAsync(cf2, Integer::sum, executor);
+        $.execute(cf1,cf2);
         int got = combine.get();
         assertEquals(2, got);
     }
@@ -273,42 +189,9 @@ public class BascomTaskFutureTest extends BaseOrchestratorTest {
         CompletableFuture<Integer> cf1 = $.task(task).ret(1);
         CompletableFuture<Integer> cf2 = $.task(task).ret(RV);
         CompletableFuture<Integer> compose = cf1.thenCompose(v -> cf2);
-        $.execute(cf2);  // Required because cf2 is otherwise invisible
+        $.execute(cf1,cf2);
         int got = compose.get();
         assertEquals(RV, got);
-    }
-
-    @Test
-    public void thenComposeAsync() throws Exception {
-        final int RV = 9;
-        CountingTask task = task(0);
-        CompletableFuture<Integer> cf1 = $.task(task).ret(1);
-        CompletableFuture<Integer> cf2 = $.task(task).ret(RV);
-        CompletableFuture<Integer> compose = cf1.thenComposeAsync(v -> cf2);
-        $.execute(cf2);  // Required because cf2 is otherwise invisible
-        int got = compose.get();
-        assertEquals(RV, got);
-    }
-
-    @Test
-    public void thenComposeAsyncExecutor() throws Exception {
-        final int RV = 9;
-        CountingTask task = task(0);
-        CompletableFuture<Integer> cf1 = $.task(task).ret(1);
-        CompletableFuture<Integer> cf2 = $.task(task).ret(RV);
-        CompletableFuture<Integer> compose = cf1.thenComposeAsync(v -> cf2, executor);
-        $.execute(cf2);  // Required because cf2 is otherwise invisible
-        int got = compose.get();
-        assertEquals(RV, got);
-    }
-
-    @Test
-    public void thenComposeCf() throws Exception {
-        CountingTask task = task(0);
-        CompletableFuture<Integer> cf1 = $.task(task).ret(2);
-        CompletableFuture<Integer> compose = cf1.thenCompose(x -> CompletableFuture.supplyAsync(() -> x * 4));
-        int got = compose.get();
-        assertEquals(8, got);
     }
 
     @Test
@@ -317,6 +200,7 @@ public class BascomTaskFutureTest extends BaseOrchestratorTest {
         CompletableFuture<Integer> cf1 = $.task(task).retHit(1);
         CompletableFuture<Integer> cf2 = $.task(task).retHit(1);
         CompletableFuture<Integer> combine = cf1.applyToEither(cf2, x -> x * 5);
+        $.execute(cf1,cf2);
         int got = combine.get();
         assertEquals(5, got);
     }
@@ -327,6 +211,7 @@ public class BascomTaskFutureTest extends BaseOrchestratorTest {
         CompletableFuture<Integer> cf1 = $.task(task).retHit(1);
         CompletableFuture<Integer> cf2 = $.task(task).retHit(1);
         CompletableFuture<Integer> combine = cf1.applyToEitherAsync(cf2, x -> x * 5);
+        $.execute(cf1,cf2);
         int got = combine.get();
         assertEquals(5, got);
     }
@@ -337,84 +222,9 @@ public class BascomTaskFutureTest extends BaseOrchestratorTest {
         CompletableFuture<Integer> cf1 = $.task(task).retHit(1);
         CompletableFuture<Integer> cf2 = $.task(task).retHit(1);
         CompletableFuture<Integer> combine = cf1.applyToEitherAsync(cf2, x -> x * 5, executor);
+        $.execute(cf1,cf2);
         int got = combine.get();
         assertEquals(5, got);
-    }
-
-    @Test
-    public void runAfterEither() throws Exception {
-        final int RV = 9;
-        CountingTask task = task(2);
-        CompletableFuture<Integer> cf1 = $.task(task).retHit(1);
-        CompletableFuture<Integer> cf2 = $.task(task).retHit(1);
-        CompletableFuture<Void> combine = cf1.runAfterEither(cf2, () -> currentTask.holder.set(RV));
-        combine.get();
-        assertEquals(RV, currentTask.holder.get());
-    }
-
-    @Test
-    public void runAfterEitherAsync() throws Exception {
-        final int RV = 9;
-        CountingTask task = task(2);
-        CompletableFuture<Integer> cf1 = $.task(task).retHit(1);
-        CompletableFuture<Integer> cf2 = $.task(task).retHit(1);
-        CompletableFuture<Void> combine = cf1.runAfterEitherAsync(cf2, () -> currentTask.holder.set(RV));
-        combine.get();
-        assertEquals(RV, currentTask.holder.get());
-    }
-
-    @Test
-    public void runAfterEitherAsyncExecutor() throws Exception {
-        final int RV = 9;
-        CountingTask task = task(2);
-        CompletableFuture<Integer> cf1 = $.task(task).retHit(1);
-        CompletableFuture<Integer> cf2 = $.task(task).retHit(1);
-        CompletableFuture<Void> combine = cf1.runAfterEitherAsync(cf2, () -> currentTask.holder.set(RV), executor);
-        combine.get();
-        assertEquals(RV, currentTask.holder.get());
-    }
-
-    @Test
-    public void runAfterBoth() throws Exception {
-        final int RV = 9;
-        CountingTask task = task(2);
-        CompletableFuture<Integer> cf1 = $.task(task).retHit(1);
-        CompletableFuture<Integer> cf2 = $.task(task).retHit(1);
-        CompletableFuture<Void> combine = cf1.runAfterBoth(cf2, () -> currentTask.holder.set(RV));
-        combine.get();
-        assertEquals(RV, currentTask.holder.get());
-    }
-
-    @Test
-    public void runAfterBothAsync() throws Exception {
-        final int RV = 9;
-        CountingTask task = task(2);
-        CompletableFuture<Integer> cf1 = $.task(task).retHit(1);
-        CompletableFuture<Integer> cf2 = $.task(task).retHit(1);
-        CompletableFuture<Void> combine = cf1.runAfterBothAsync(cf2, () -> currentTask.holder.set(RV));
-        combine.get();
-        assertEquals(RV, currentTask.holder.get());
-    }
-
-    @Test
-    public void runAfterBothAsyncExecutor() throws Exception {
-        final int RV = 9;
-        CountingTask task = task(2);
-        CompletableFuture<Integer> cf1 = $.task(task).retHit(1);
-        CompletableFuture<Integer> cf2 = $.task(task).retHit(1);
-        CompletableFuture<Void> combine = cf1.runAfterBothAsync(cf2, () -> currentTask.holder.set(RV), executor);
-        combine.get();
-        assertEquals(RV, currentTask.holder.get());
-    }
-
-    @Test
-    public void acceptEither() throws Exception {
-        CountingTask task = task(2);
-        CompletableFuture<Integer> cf1 = $.task(task).retHit(1);
-        CompletableFuture<Integer> cf2 = $.task(task).retHit(1);
-        CompletableFuture<Void> either = cf1.acceptEither(cf2, x -> currentTask.holder.set(x));
-        either.get();
-        assertEquals(1, currentTask.holder.get());
     }
 
     @Test
@@ -423,17 +233,8 @@ public class BascomTaskFutureTest extends BaseOrchestratorTest {
         CompletableFuture<Integer> cf1 = $.task(task).retHit(1);
         CompletableFuture<Integer> cf2 = $.task(task).retHit(1);
         CompletableFuture<Void> either = cf1.acceptEitherAsync(cf2, x -> currentTask.holder.set(x));
-        either.get();
-        assertEquals(1, currentTask.holder.get());
-    }
-
-    @Test
-    public void acceptEitherAsyncExecutor() throws Exception {
-        CountingTask task = task(2);
-        CompletableFuture<Integer> cf1 = $.task(task).retHit(1);
-        CompletableFuture<Integer> cf2 = $.task(task).retHit(1);
-        CompletableFuture<Void> either = cf1.acceptEitherAsync(cf2, x -> currentTask.holder.set(x), executor);
-        either.get();
+        $.executeAndWait(cf1,cf2);
+        sleep(20);
         assertEquals(1, currentTask.holder.get());
     }
 
@@ -441,6 +242,7 @@ public class BascomTaskFutureTest extends BaseOrchestratorTest {
     public void whenComplete() throws Exception {
         int expectedValue = 567;
         CompletableFuture<Integer> cf1 = $.task(task(1)).retHit(expectedValue);
+        cf1.join();
         int got = cf1.whenComplete((v, ex) -> verify(expectedValue, v, ex)).get();
         assertEquals(expectedValue,got);
     }
@@ -449,54 +251,8 @@ public class BascomTaskFutureTest extends BaseOrchestratorTest {
     public void whenCompleteAsync() throws Exception {
         int expectedValue = 567;
         CompletableFuture<Integer> cf1 = $.task(task(1)).retHit(expectedValue);
+        cf1.join();
         int got = cf1.whenCompleteAsync((v, ex) -> verify(expectedValue, v, ex)).get();
         assertEquals(expectedValue,got);
     }
-
-    @Test
-    public void whenCompleteAsyncExecutor() throws Exception {
-        int expectedValue = 567;
-        CompletableFuture<Integer> cf1 = $.task(task(1)).retHit(expectedValue);
-        int got = cf1.whenCompleteAsync((v, ex) -> verify(expectedValue, v, ex),executor).get();
-        assertEquals(expectedValue,got);
-    }
-
-    @Test
-    public void handle() throws Exception {
-        final int HV = 11;
-        CompletableFuture<String> cf1 = $.task(task(0)).rets("no_matter");
-        CompletableFuture<Integer> cf2 = $.task(task(0)).faults(cf1).handle((x, y) -> HV);
-        int got = cf2.get();
-        assertEquals(HV, got);
-    }
-
-    @Test
-    public void handleAsync() throws Exception {
-        final int HV = 11;
-        CompletableFuture<String> cf1 = $.task(task(0)).rets("no_matter");
-        CompletableFuture<Integer> cf2 = $.task(task(0)).faults(cf1).handleAsync((x, y) -> HV);
-        int got = cf2.get();
-        assertEquals(HV, got);
-    }
-
-    @Test
-    public void handleAsyncExecutor() throws Exception {
-        final int HV = 11;
-        CompletableFuture<String> cf1 = $.task(task(0)).rets("no_matter");
-        CompletableFuture<Integer> cf2 = $.task(task(0)).faults(cf1).handleAsync((x, y) -> HV, executor);
-        int got = cf2.get();
-        assertEquals(HV, got);
-    }
-
-
-    @Test
-    public void exceptionally() throws Exception {
-        final int HV = 999;
-        CompletableFuture<Integer> cf1 = $.task(task(1)).retHit(1);
-        CompletableFuture<Integer> cf2 = $.task(task(0)).fault(cf1).exceptionally(e -> HV);
-        int got = cf2.get();
-        assertEquals(HV, got);
-    }
-
-
 }